First step in the rearchitecture is to move some of the clicast command stuff to a new python script ...
